fotory/pyproject.toml
[tool.poetry]
name = "fotory"
version = "0.1.0"
description = "Fotory - digital album"
authors = ["Qiyas <107604130+qiyascc@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.3"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
